backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymupdf
anthropic
voyageai
pinecone-client
langchain
python-multipart
requests
numpy
pandas
faiss-cpu
flask
flask-cors
# Keeping data processing libs from earlier tasks
tiktoken
